use alternative DataDog Python package This python package, 'datadog', has better documentation and seems simpler to use then the prior one 'datadog_api_client'.
